+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+* `ElectrumxD`: assert Python is available at runtime
 * Fix naming inconsistencies
 * Bump `bitcoin_hashes` to v1.1.0
 * Add and apply `rustfmt.toml`